Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 void main()
 {
-    int a,b,c,Average;
-    scanf("%d%d%d",&a,&b,&c);
+    float a,b,c,Average;
+    scanf("%f%f%f",&a,&b,&c);
     Average=(a+b+c)/3;
     printf("Average: %.2f",Average);
 }
